backend/requirements.txt
# FastAPI Backend Dependencies

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.19
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Vector Store & Embeddings
chromadb>=0.5.23
openai>=1.60.0
google-generativeai>=0.8.4
anthropic>=0.42.0
sentence-transformers>=2.2.0

# Document Processing
pymupdf>=1.25.0
python-docx>=1.1.0

# Web Search
requests>=2.32.0

# Utilities
python-dotenv>=1.0.1
uuid>=1.30
